# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tourist Route Management API",
        "description": "API documentation for the Tourist Route project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/dishes": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "List dishes",
                "operationId": "ed30e3af81982de8e2785760bd0b5e9f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Create dish",
                "operationId": "71548f02737f7500c547d06dbf29bf1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "type": "number",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/dishes/{dishes}": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Show dish",
                "operationId": "975c4c54877c453ccc18ecbe5c4fcf24",
                "parameters": [
                    {
                        "name": "dishes",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Update dish",
                "operationId": "a09b540195bd227dd391e83e882bb7ce",
                "parameters": [
                    {
                        "name": "dishes",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "price": {
                                        "type": "number",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Delete dish",
                "operationId": "d90720fa83ea7cee2d3e05f717cf681a",
                "parameters": [
                    {
                        "name": "dishes",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/distinations": {
            "get": {
                "tags": [
                    "Distination"
                ],
                "summary": "List distinations",
                "operationId": "1f2accb557628af5db09240d52197d6e",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Distination"
                ],
                "summary": "Create distination",
                "operationId": "a81c4169a01431d01cf1bee6974cca98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "accommodation",
                                    "dishes",
                                    "places",
                                    "activities"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "accommodation": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "dishes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "places": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/distinations/{distination}": {
            "get": {
                "tags": [
                    "Distination"
                ],
                "summary": "Show distination",
                "operationId": "3716f47a5a6a2865e6bfe6d3370f2b84",
                "parameters": [
                    {
                        "name": "distination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Distination"
                ],
                "summary": "Update distination",
                "operationId": "17b19f8a25a136799f5b906a64fd5a43",
                "parameters": [
                    {
                        "name": "distination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "accommodation": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Distination"
                ],
                "summary": "Delete distination",
                "operationId": "c6a4ed78def6f1fcf17aa6179c3881f6",
                "parameters": [
                    {
                        "name": "distination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}